requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
airflow-provider-duckdb
duckcli
dbt-duckdb
dbt-core
pandas
pyarrow
